controller: allow kubeconfig overrides
In `spec.secretRef`, if `kubeconfig` is already specified, then we should just
honor it and use it as the kubeconfig going forward. This allows a use case for
cross-cluster deployments. In that use case, all the release objects can be
specified from one single cluster, but managing all the other clusters via
kubeconfig overrides in secretRef. This is a rare case, but might still be
useful sometimes.